requirements.txt
opencv-python-headless
streamlit
pandas
numpy
pillow
